social_media_api/requirements.txt
Django>=5.2.4
djangorestframework>=3.14.0
Pillow>=10.0.0
requests>=2.31.0
